+++ src/test/java/com/backendoori/ootw/user/controller/UserControllerTest.java
@@ -10,10 +10,10 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 
 import java.time.LocalDateTime;
 import java.time.temporal.ChronoUnit;
-import com.backendoori.ootw.config.jwt.TokenProvider;
 import com.backendoori.ootw.exception.AlreadyExistEmailException;
 import com.backendoori.ootw.exception.IncorrectPasswordException;
 import com.backendoori.ootw.exception.NotExistUserException;
+import com.backendoori.ootw.security.jwt.TokenProvider;
 import com.backendoori.ootw.user.dto.LoginDto;
 import com.backendoori.ootw.user.dto.SignupDto;
 import com.backendoori.ootw.user.dto.TokenDto;
